# api/api.yaml
openapi: 3.0.0
info:
  title: ZidiBackend API
  version: 1.0.0
  description: API for managing organizations and users in ZidiBackend
paths:
  /organizations:
    post:
      summary: Create an organization
      operationId: CreateOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: Get all organizations
      operationId: GetOrganizations
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Number of organizations to return (pagination)
          required: false
          schema:
            type: integer
            minimum: 1
            example: 10
            default: 10
        - name: offset
          in: query
          description: Number of organizations to skip before starting to return results
          required: false
          schema:
            type: integer
            minimum: 0
            example: 0
            default: 0
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /organizations/{organization_id}:
    get:
      summary: Get an organization by ID
      operationId: GetOrganizationById
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update an organization
      operationId: UpdateOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete an organization
      operationId: DeleteOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Organization deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
...
  /organizations/search:
    get:
      summary: Get an organization by name
      operationId: GetOrganizationByName
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Number of organizations to return (pagination)
          required: false
          schema:
            type: integer
            minimum: 1
            example: 10
        - name: offset
          in: query
          description: Number of organizations to skip before starting to return results
          required: false
          schema:
            type: integer
            minimum: 0
            example: 0
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
  /users:
    post:
      summary: Create a user
      operationId: CreateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequestBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"

    get:
      summary: Get all users
      operationId: GetUsers
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Number of users to return (pagination)
          required: false
          schema:
            type: integer
            minimum: 1
            example: 10
        - name: offset
          in: query
          description: Number of users to skip before starting to return results
          required: false
          schema:
            type: integer
            minimum: 0
            example: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /payments:
    post:
      summary: Create a new payment
      operationId: createPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
      responses:
        "201":
          description: Payment successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid request data
        "500":
          description: Internal server error

    get:
      summary: Get all payments
      operationId: getAllPayments
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            example: 0
      responses:
        "200":
          description: A list of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  payments:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of payments in the database.
                      payments:
                        type: array
                        items:
                          $ref: "#/components/schemas/Payment"
        "500":
          description: Internal server error
  /payments/{id}:
    get:
      summary: Get a payment by ID
      operationId: getPaymentById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Payment not found
        "500":
          description: Internal server error

    put:
      summary: Update a payment by ID
      operationId: updatePaymentById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
      responses:
        "200":
          description: Payment updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid request data
        "404":
          description: Payment not found
        "500":
          description: Internal server error
  /payments/organization/{organization_id}:
    get:
      summary: Get payments by organization ID
      operationId: getPaymentsByOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            example: 0
      responses:
        "200":
          description: List of payments for an organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  payments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Payment"
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps with documentation
  schemas:
    Organization:
      type: object
      required:
        - id
        - email
        - contact_person_name
        - contact_person_phone
        - address
        - industry
        - company_size
        - company_name
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        contact_person_name:
          type: string
        contact_person_phone:
          type: string
        address:
          type: string
        industry:
          type: string
        company_size:
          type: integer
        company_name:
          type: string
...
    User:
      type: object
      required:
        - id
        - firstname
        - lastname
        - email
        - address
        - password
        - role
        - organization_id
      properties:
        id:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
    Payment:
      type: object
      required:
        - id
        - organization_id
        - campaign_id
        - amount
        - transaction_ref
        - transaction_id
        - status
      properties:
        id:
          type: string
          format: uuid
          example: "5c2a4d94-27f8-4fbd-8f74-8dd8c7c8e49c"
        organization_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        campaign_id:
          type: string
          format: uuid
          example: "2d6f2952-9a5b-4663-91e1-67d573a3f940"
        amount:
          type: number
          format: float
          example: 100.50
        transaction_ref:
          type: string
          example: "TXN123456789"
        transaction_id:
          type: string
          example: "543456789654789654"
        status:
          type: string
          enum: [pending, completed, failed]
          example: "pending"
...
    CreateUserRequestBody:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - address
        - role
        - organization_id
        - password
        - confirmPassword
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
          format: password
        address:
          type: string
        confirmPassword:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
...
    Error:
      type: object
      required:
        - message
        - error_code
      properties:
        error_code:
          type: string
        message:
          type: string
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []